Issue #2, spin-off some of the file/directory methods of AipDownloader. Reducing the size of the class and hopefully exposing the logic better.
